third_party/WebKit/LayoutTests/webaudio/internals/audiocontext-lock-threading-race-expected.txt

CONSOLE WARNING: line 14: The Web Audio autoplay policy will be re-enabled in Chrome 70 (October 2018). Please check that your website is compatible with it. https://goo.gl/7K7WLu
TSan shouldn't complain about threading races of AudioContext::lock().

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
